+++ p2ploanplatform/src/main/java/com/scut/p2ploanplatform/entity/CreditInfo.java
@@ -2,6 +2,8 @@ package com.scut.p2ploanplatform.entity;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * @author: Light
  * @date: 2019/6/17 10:59
@@ -19,17 +21,17 @@ public class CreditInfo {
     /**
      * 收入（单位:人民币）
      */
-    private Double income;
+    private BigDecimal income;
 
     /**
      * 家庭收入 （单位:人民币）
      */
-    private Double familyIncome;
+    private BigDecimal familyIncome;
 
     /**
      *资产 （单位:人民币）
      */
-    private Double assets;
+    private BigDecimal assets;
 
     /**
      * 家庭成员人数
@@ -39,7 +41,7 @@ public class CreditInfo {
     /**
      * 负载 （单位:人民币）
      */
-    private Double debt;
+    private BigDecimal debt;
 
     /**
      * 信用分数
